requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pypdf==4.2.0
python-docx==1.1.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
langchain==0.2.5
ollama==0.2.1